Update DDL to v3 (add FK to account_roles)
NOTE: The auto-generated ddl files didn't work for MySQL because
the roles table was MyISAM still and needed to be InnoDB before
the account_roles FK constraints could be added. Fortunately the
fix was as simple as moving the ALTER TABLE roles line to the top
of the file ddl/MySQL/deploy/3/001-auto.sql

To test the upgrade multiple times during debugging the error, I
followed this procdure.

Delete the database in the appropriate way
(DROP DATABASE rpgcat_db in MySQL; rm rpgcat.db with SQLite)

Then install version 2 schema (-c specifies connection string)

./script/rpgcat_dh.pl -s RPGCat::Schema -I lib -o ./ddl/ \
    -c 'dbi:mysql:rpgcat_db;user=rpgcat_user;password=rpgcat_pass' \
    install --target 2

Once that's done, restart the application server and it should
upgrade the db schema to v3 (or die with a horrible error)