A good start to #15 - still needs some work
This is the email sending code - EMKit - that uses Email::MIME::Kit
and generates emails based on the content of templates/email/

There is no password recovery page yet, but if you don't have an
account it informs you via email as per guidelines that I found
that prevent info leakage.

"sorry, you don't have an account" in a webpage is an instant
'try a different email' flag to an attacker so instead we say
"we've emailed instructions" and then say in the email that
they don't have an account and to register a new one.