Added link to map page to character list
If you choose to play a character, it logs you into that one and
redirects to the map page. Currently the map page has no content
except a witty comment but that's coming next.

The NeedsCharacter role should handle blocking access to any
controllers that specifically need an active character, just
like the NeedsAccount role blocks access to anything that needs
a valid account to be logged in.